requirements.txt
pillow
tk
pytest
